Keybindings, Pointer-Problem
Well, in general it works, but there's one problem:
The TUI struct, needs a node-pointer which points to one of its attributes:
"self.keybindings". Since it's not possible to create an pointer on a attribute
on the same struct, we'd have to wrap each Vector into an Rc. I think the
HashMap approach is better than using vectors, since they don't provide an
advantage anymore.

In other words, this is the Problem:

src/tui/model.rs

        Ok(Tui {
            sidebar,
            maillist,
            connections: Vec::new(),
            config: config,
            keybindings: keybindings_var,
            // here's the problem
            keybinding_node: &Self { keybindings: keybinding_var, .. },
            need_redraw: true,
            run: true,
        })