Mail/Msg refactor suggestion
This commit should provide a suggestion how to "interact" with the mails in
himalaya.

I didn't wrote a full documentation yet since I don't know if this is gonna be
accepted or not. The code isn't that big, so I think that it's understandable,
if not, feel free to ask!

Since himalaya is mainly a Cli-Mail-Client it might be better to add a
possibility to add a something like a vector into the `new()` function of `Mail`
which includes all parts of `Mail` which has to be loaded. For example only the
`attachments` attribute of `Mail` which could reduce the load time (although it
might be just some microseconds).