CLI Refactor
Moved all cli.rs to the src/cli directory.
This makes it possible to call the subcmds() and matches() function according
to the filename like `cli::mbox::subcmds()` instead of `mbox_subcmds()` (not
really a big deal but it's a suggestions).

Also added some documentation how the `src/cli` directory works.