Add man page
With the inclusion of himalaya to various Linux distributions we should
fulfill the requirement of having a manual page for each executable.

I chose to write the man page in markdown for easier editing.

It can then be transformed into a regular man page via pandoc:
`pandoc --standalone --to man himalaya.1.md -o himalaya.1`